% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_utils.R
\name{adjust_stg_z}
\alias{adjust_stg_z}
\title{Use a calibration STG file to adjust the Z coordinates of a "stage_coords" data frame}
\usage{
adjust_stg_z(
  stage_coords,
  calib_stg_path = system.file("stage_positions-calib_tilt_B2_top_left.STG", package =
    "rcell2.cellid"),
  origin_at_corner = "top-right",
  plot_error = FALSE
)
}
\arguments{
\item{stage_coords}{Stage coordinates data frame, as produced by "make_stage_list".}

\item{calib_stg_path}{Path to an STG file with calibration points. Their origin must be the same as the origin used in \code{stage_coords}.}

\item{origin_at_corner}{Corner of the well at which the coordinate origin was set.}
}
\description{
Use a calibration STG file to adjust the Z coordinates of a "stage_coords" data frame
}
