% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{arguments_summary}
\alias{arguments_summary}
\title{Print rcellid arguments summaries}
\usage{
arguments_summary(arguments)
}
\arguments{
\item{arguments}{The "arguments" dataframe, output from \code{rcell2.cellid::arguments()}.}
}
\description{
A function to print some summaries, to check cellArgs2 output.
}
