% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{arguments_to_images}
\alias{arguments_to_images}
\title{Make and "images" dataframe from "arguments" dataframe}
\usage{
arguments_to_images(arguments, bf_as_fl = F)
}
\arguments{
\item{arguments}{The "arguments" data frame, as generated by \code{rcell2.cellid::arguments()}.}

\item{bf_as_fl}{Set this to TRUE if the input already contains BF images as channels (i.e. if \code{bf_as_fl} was set to TRUE when generating the arguments data frame).}
}
\value{
A data.frame similar to \code{get_cell_data()$images}.
}
\description{
The images dataframe is needed by many rcell2 functions. If it is not available from the output of \code{load_cell_data} or \code{get_cell_data}, then this function can help.
}
\details{
It essentially does a pivot_longer of the arguments.
}
