% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cell.load.alt}
\alias{cell.load.alt}
\title{Cargar el output de cell-id}
\usage{
cell.load.alt(
  path,
  pdata = NULL,
  position.pattern = ".*Position(\\\\d+).*",
  fluorescence.pattern = "^(BF|[GCYRT]FP|[GCYRT]\\\\d+)_Position\\\\d+.*.tif$",
  ucid.zero.pad = 4,
  append.posfix = NULL,
  ...
)
}
\arguments{
\item{path}{Path to CellID's output directory, tipically also the images directory.}

\item{pdata}{Path to metadata CSV file.}

\item{position.pattern}{Regex describing what the position string looks like (default ".*Position(\\d+).*") including a capturing group for the position ID number (coerced to integer).}

\item{fluorescence.pattern}{Regex describing what the fluorescence/channel ID string looks like (default "^([GCYRT]FP|[GCYRT]\\d+)_Position\\d+_time\\d+.tif$"). There must be only one capturing group, ant it must be for the channel identifier.}

\item{ucid.zero.pad}{Amount of decimal digits for the cellID (defaults 4, corresponding to a maximum of 9.999 cellIDs and 9999 positions).}

\item{append.posfix}{String appended to the channel ID extracted by `fluorescence.pattern` (`NULL` by default, but "FP" is usual).}

\item{...}{
  Arguments passed on to \code{\link[=load_out_all]{load_out_all}}
  \describe{
    \item{\code{out_file_pattern}}{Regex matching CellID's main output file.}
    \item{\code{out_mapping_pattern}}{Regex matching CellID's image mapping output file.}
    \item{\code{fix_id_columns}}{Whether apply the \code{fix_id_columns_fun} summary function to problematic ID columns (with non-unique vales per cell across channels).}
    \item{\code{fix_id_columns_fun}}{Summary function for \code{fix_id_columns}, uses dplyr's \code{first} by default (thereby using the value of the first channel).}
  }}
}
\value{
A list of dataframes: data (CellID data), images (images metadata and paths), image_maping (extra mapping metadata from CellID: BF to FL correspondence, channel flag, bf_as_fl flag, and one-letter channel encoding).
}
\description{
Cargar el output de cell-id
}
