% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cell2}
\alias{cell2}
\title{Function to run CellID}
\usage{
cell2(
  arguments,
  cell.command = NULL,
  n_cores = NULL,
  debug_flag = 0,
  dry = F,
  encode_cellID_in_pixels = F,
  fill_interior_pixels = F,
  label_cells_in_bf = F,
  output_coords_to_tsv = F,
  save.logs = T,
  verbose = T,
  progress = F,
  check_fail = F
)
}
\arguments{
\item{arguments}{An argument data.frame, as built by \code{rcell2.cellid::arguments}.}

\item{cell.command}{By default \code{NULL}, to use the built-in binary. Otherwise a path to a CellID binary executable (get if from https://github.com/darksideoftheshmoo/cellID-linux).}

\item{n_cores}{Number of cores to use for position-wise parallelization,internally capped to number of positions in \code{arguments}. Set to 1 to disable parallelization. If NULL, defaults to available cores - 1.}

\item{debug_flag}{Set to 0 to disable CellID printf messages (builtin CellID only).}

\item{dry}{Do everything without actually running CellID, print the commands that would have been issued.}

\item{encode_cellID_in_pixels}{Set to TRUE to write cell interior and boundary pixels with intensity-encoded CellIDs and blank the rest of the image (CellID option '-s').}

\item{fill_interior_pixels}{Set to TRUE to fill each cell interior area in the output image file with intensity-labeled pixels (CellID option '-i').}

\item{label_cells_in_bf}{Set to TRUE to enable labeling of cells with their CellID in the BF output image using number characters (CellID option '-l', default FALSE).}

\item{output_coords_to_tsv}{Set to TRUE to write cell interior masks and boundary pixels data to a .tsv file in the output directory (CellID option '-m').}

\item{save.logs}{Set to TRUE to save CellID logs to text files, into the output directory of their corresponding position.}

\item{verbose}{Print startup messages.}

\item{progress}{Print a progress bar. Requires the \code{doSNOW} package.}

\item{check_fail}{Whether to produce an error if any output file or directory already exists.}
}
\value{
A dataframe with one column indicating the issued commands and exit codes (in the command.output column). If the execution was sucessful, now you may run \code{rcell2::load_cell_data} or \code{rcell2.cellid::cell.load.alt} to get the results from the CellID output, typically located at the images path.
}
\description{
Function to run CellID
}
