% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cellid_output_descriptions}
\alias{cellid_output_descriptions}
\title{Cell-ID output descriptions}
\usage{
cellid_output_descriptions(list.output = T)
}
\arguments{
\item{list.output}{Return the descriptions as a named list (TRUE), or as a data.frame (FALSE).}
}
\description{
Cell-ID output descriptions
}
