% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cero_a_la_izquierda}
\alias{cero_a_la_izquierda}
\title{Pad numbers to the same length with leading zeros}
\usage{
cero_a_la_izquierda(x, pad_char = "0")
}
\description{
Pad numbers to the same length with leading zeros
}
