% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiff_info.R
\name{get_plane_info}
\alias{get_plane_info}
\title{Get stage plane information from MetaMorph images}
\usage{
get_plane_info(images)
}
\arguments{
\item{images}{The images dataframe as loaded from Cell-ID's output by \code{get_cell_data}.}
}
\description{
Get stage plane information from MetaMorph images
}
