% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_utils.R
\name{get_spos_template}
\alias{get_spos_template}
\title{A function to copy the spos XLSX spreadsheet template to the current working directory.}
\usage{
get_spos_template(file_name = "stage_positions.xlsx")
}
\arguments{
\item{file_name}{File name for the downloaded workflow template.}
}
\description{
The template file is included in \code{cell2.cellid}.
}
