% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiff_info.R
\name{get_time_info}
\alias{get_time_info}
\title{Load acquisition time information from MetaMorph's TIFF files}
\usage{
get_time_info(
  image_list = NULL,
  images = NULL,
  channels = "BF",
  join_to_images = T
)
}
\arguments{
\item{image_list}{The output from the \code{arguments} function. Consider filtering to include only the first position.}

\item{images}{The images dataframe as loaded from Cell-ID's output by \code{get_cell_data}.}

\item{channels}{Vector of channel IDs (e.g. "BF").}

\item{join_to_images}{If TRUE, the position/frame metadata form the images datafram will be joined to the extracted acquisition times.}
}
\description{
Time information is loaded from the XML inserted by MetaMorph into the TIFF files' metadata.
}
\details{
Time is loaded as POSIXlt and POSIXct into different columns (the 'ct' suffix indicates POSIXct). See \code{DateTimeClasses} for details.
}
