% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_macros.R
\name{ijm_open_hyperstack}
\alias{ijm_open_hyperstack}
\title{Make an ImageJ macro to browse the images before segmentation}
\usage{
ijm_open_hyperstack(
  images,
  use_out = 0:1,
  temp_dir_name = "ijm_set",
  macro_file = TRUE,
  fix_order = TRUE,
  add_tools = TRUE
)
}
\arguments{
\item{images}{The "images" dataframe, as loaded by get_cell_data, load_cell_data, or as generated by arguments_to_images from arguments.}

\item{use_out}{Wether to use original images (FALSE/0), the segmented output (TRUE/1), or both (0:1).}

\item{temp_dir_name}{Base-name for the temporary directory where renamed links to images are stored. It will be created as a subdirectory of \code{tempdir()}. Any directory in this path with the same name will be deleted before use.}

\item{macro_file}{Write the macro to temporary file/directory. Else, a path to the file where to save the macro.}

\item{fix_order}{Symlink images to a temporary directory with new file names. The new sorting order will make images appear with the correct channel-slice-frame mapping in ImageJ.}

\item{add_tools}{Add commands to the macro that open handy tools (e.g. for tagging events in single cells, adjusting brightness, etc.).}
}
\description{
Due to an order restriction on ImageJ Virtual Hyperstacks, t.frame is mapped to "Channel", position is mapped to "Slice", and channel is mapped to "frames".
}
\details{
The function writes a macro to a temp
}
