% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_macros.R
\name{ijm_open_segmentation}
\alias{ijm_open_segmentation}
\title{IJ macro to open .out files as a virtual hyperstack}
\usage{
ijm_open_segmentation(cellid.args)
}
\arguments{
\item{cellid.args}{The "arguments" dataframe, as produced by \code{rcell2.cellid::arguments}.}
}
\description{
IJ macro to open .out files as a virtual hyperstack
}
