% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_macros.R
\name{imagej_fft_filter}
\alias{imagej_fft_filter}
\title{Run ImageJ FFT filter macro from R}
\usage{
imagej_fft_filter(
  pic.path,
  script.path = system.file(file.path("imagej_macros", "FFT_filter_on_BFs_R.txt"),
    package = "rcell2.cellid"),
  ...
)
}
\arguments{
\item{pic.path}{Path to the directory containing the image files (passed as extra.args to imagej_macro_run).}

\item{script.path}{Path to the ImageJ macro. Defaults to built-in macro.}

\item{...}{
  Arguments passed on to \code{\link[=imagej_macro_run]{imagej_macro_run}}
  \describe{
    \item{\code{imagej.path}}{Path to the ImageJ binary (a path to "ImageJ-linux64" or equivalent).}
    \item{\code{extra.args}}{A string with extra arguments to the ImageJ command, pasted at the end.}
    \item{\code{headless}}{Wether ImageJ should be run headlessly (no GUI).}
    \item{\code{wait}}{a logical (not \code{NA}) indicating whether the \R
    interpreter should wait for the command to finish, or run it
    asynchronously.  This will be ignored (and the interpreter will
    always wait) if \code{intern = TRUE}. When running the command
    asynchronously, no output will be displayed on the \code{Rgui}
    console in Windows (it will be dropped, instead).}
  }}
}
\description{
Passes an ImageJ FFT filter on files matching \code{BF.*.tif} in the target directory.
}
\details{
The modified images are saved to a new subdirectory with default name: "filtered/" (hardcoded in the macro).

Run \code{???imagej_macro_run} for more info.
}
