% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{join_mdas}
\alias{join_mdas}
\title{Generate metadata for renaming multiple MDA runs into one set}
\usage{
join_mdas(index_pattern = "^MDA(\\\\d+)_.*", ..., print_plot = T)
}
\arguments{
\item{index_pattern}{Regular expression with a single capturing group for the MDA group index. The index matched by the pattern must be an integer, and mus uniquely identify each MDA run.}

\item{...}{
  Arguments passed on to \code{\link[=rename_mda]{rename_mda}}
  \describe{
    \item{\code{images.path}}{Path to the directory containing the original images. Can be NULL if \code{file.names} is provided.}
    \item{\code{rename.path}}{Path to the target directory. If \code{NULL} (the default) images are sent to a new "renamed" sub directory of \code{images.path}. The directory will be created if it does not exist.}
    \item{\code{rename.dataframe}}{If \code{TRUE}, no renaming takes place and a data frame is returned instead. It can be altered by the user, and then passed to this same argument to rename the images with the adjusted names or mappings. The data frame is normally returned in the "status" item of the regular output.}
    \item{\code{rename.function}}{Either \code{\link[base]{file.copy}}, \link[base]{file.symlink} or a similar function. Set to \code{NULL} to disable renaming (i.e. for testing purposes).}
    \item{\code{identifier.pattern}}{Regex defining the iamge file pattern, with gropus for identifier in the file names.}
    \item{\code{identifier.info}}{Character vector with strings "pos", "t.frame", and "ch" (channel), in the same order in which they appear in the \code{identifier.pattern}. If an element in the vector is named, the name is prefixed to the identifier in the final file name (for example, by default, "Position" is prepended to the position number; but channel has no prefix).}
    \item{\code{channel.maping.df}}{A dataframe with two columns: "ch" holding the original channel names in the source files (coerced to character), and "ch.name" with the new names for each channel.}
    \item{\code{file.ext}}{File extension to use in the final file name, such as: ".tif".}
    \item{\code{skip.thumbs.pat}}{A regex pattern to filter out files. Convenient if the MDA output thumbnails for each image. Set to \code{NULL} to disable.}
    \item{\code{cleanup.first}}{Set to \code{TRUE} to remove all files within the \code{rename.path} directory before renaming. \code{FALSE} by default.}
    \item{\code{file.names}}{A character vector of image names to be processed; as an alternative to listing files in \code{images.path}. Ignored if \code{images.path} is provided.}
  }}

\item{print_plot}{Prints a plot that can help check visually if the final image set will be complete.}
}
\value{
A data frame to be passed to the \code{rename.dataframe} argument of \code{rename_mda}.
}
\description{
The output of this function is meant to be passed on to \code{rename_mda}.
}
\details{
Find details in the help page of \code{rename_mda}.

Example:
data.dir <- "/path/to/images/"
rename.df <- rename_mdas(
  images.path = data.dir,
  index_pattern = "^MDA(\\d+)_.*",
  # Example pattenrs:
  identifier.pattern=".*_w(\\d).*_s(\\d{1,2})_t(\\d{1,2}).TIF$",
  identifier.info = c("ch", Position="pos", time="t.frame"),
  channel.maping.df = data.frame(ch = 1:4, 
  ch.name = c("TFP", "BF", "RFP", "YFP")))
result <- rename_mda(
  images.path = data.dir, 
  rename.dataframe = rename.df,
  channel.maping.df = data.frame(ch = 1:4, ch.name = c("TFP", "BF", "RFP", "YFP")))
}
