% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{load_out_all}
\alias{load_out_all}
\title{Una función para leer y combinar todos los archivos "out".}
\usage{
load_out_all(
  path,
  position.pattern = ".*Position(\\\\d+)$",
  out_file_pattern = "^out_all$",
  out_mapping_pattern = "^out_bf_fl_mapping$",
  fluorescence.pattern = ".*(BF|[A-Z]FP)_Position.*",
  fix_id_columns = F,
  fix_id_columns_fun = dplyr::first
)
}
\arguments{
\item{path}{Path to CellID's output directory, tipically also the images directory.}

\item{position.pattern}{Regex describing what the position string looks like (default ".*Position(\\d+).*") including a capturing group for the position ID number (coerced to integer).}

\item{out_file_pattern}{Regex matching CellID's main output file.}

\item{out_mapping_pattern}{Regex matching CellID's image mapping output file.}

\item{fluorescence.pattern}{Regex describing what the fluorescence/channel ID string looks like (default "^([GCYRT]FP|[GCYRT]\\d+)_Position\\d+_time\\d+.tif$"). There must be only one capturing group, ant it must be for the channel identifier.}

\item{fix_id_columns}{Whether apply the \code{fix_id_columns_fun} summary function to problematic ID columns (with non-unique vales per cell across channels).}

\item{fix_id_columns_fun}{Summary function for \code{fix_id_columns}, uses dplyr's \code{first} by default (thereby using the value of the first channel).}
}
\value{
A list of two dataframes: `d` contains the actual output, and `out.map` contains image paths and metadata.
}
\description{
Una función para leer y combinar todos los archivos "out".
}
\keyword{internal}
