% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_parameter_scan.R
\name{make_info_box}
\alias{make_info_box}
\title{Prepare information text box to annotate an image with it's segmentation parameters and metadata}
\usage{
make_info_box(images)
}
\arguments{
\item{images}{The "results.bound" data frame produced by \code{parameter_scan}, filtered to contain only one channel.}
}
\value{
A character vector for an information "text box", with each line padded to square length.
}
\description{
Internal use.
}
