% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_utils.R
\name{make_pdata}
\alias{make_pdata}
\title{Prepare a pdata dataframe from a template spreadsheet}
\usage{
make_pdata(
  spreadsheet_path,
  pdata_sheets = c(),
  plot_metadata = TRUE,
  write_pdata = TRUE
)
}
\arguments{
\item{spreadsheet_path}{Path to the an XLSX spreadsheet (e.g. "stage_positions.xlsx") with the following sheets: "pdata", "well_order", "well_images." A template file can be obtained with \code{get_spos_template}.}

\item{pdata_sheets}{Extract additional metadata from these sheets in the template spreadsheet file.}

\item{plot_metadata}{If TRUE, plot the extracted metadata.}

\item{write_pdata}{If TRUE, write a "pdata.csv" file next to the original template file.}
}
\description{
Extract metadata information from a template spreadsheet (e.g. as the one provided by \code{get_spos_template}), joining it to the "pdata" table in that spreadsheet.
}
