% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_utils.R
\name{make_stage_list}
\alias{make_stage_list}
\title{Generate stage position data for a 384 well-plate experiment}
\usage{
make_stage_list(
  spreadsheet_path,
  stg_output_path = "stage_list.STG",
  well_sep = 4500,
  well_width = 3300,
  fov_width = 500,
  origin_at_corner = "top-right",
  origin_at_pos = 1,
  af_offset_default = 0,
  z_default = 0
)
}
\arguments{
\item{spreadsheet_path}{Path where the STG file will be saved. This file is meant to be imported with MetaMorph's MDA.}

\item{well_sep}{Center-to-center distance between wells in the plate (in microns).}

\item{well_width}{Width of the bottom of the wells (in microns).}

\item{fov_width}{Width of the field of view of the final image (in microns).}

\item{origin_at_corner}{Corner of the well at which the coordinate origin was set.}

\item{origin_at_pos}{Stage position that contains the origin (at the corner selected by \code{origin_at_corner}).}

\item{af_offset_default}{Default value for the hardware auto-focus offset.}

\item{z_default}{Default value for the Z-coordinate (i.e. objective lens height).}
}
\description{
Generate stage position data for a 384 well-plate experiment
}
\details{
This helper function generates a ".STG" file that can be imported into MetaMorph's MDA, to load stage positions for an experiment.

The function requires an XLSX spreadsheet as an input, which must contain sheets with:

1. "pdata" with the position metadata,

2. "well_order" with a representation of the plate indicating the order in which to take pictures,

3. and "well_images" with a representation of the plate indicating how many images to acquire in each well.

A template XLSX file can be obtained with \code{get_spos_template}.
}
