% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_utils.R
\name{make_wells}
\alias{make_wells}
\title{Extract well order and image numbers from the template spreadsheet}
\usage{
make_wells(
  spreadsheet_path,
  well_order_sheet = "well_order",
  well_images_sheet = "well_images"
)
}
\description{
Extract well order and image numbers from the template spreadsheet
}
