% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_utils.R
\name{model_stg_z}
\alias{model_stg_z}
\title{Fit a linear model to the XY coordinates and predict Z}
\usage{
model_stg_z(calib_stg_list)
}
\description{
Fit a linear model to the XY coordinates and predict Z
}
