% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\docType{data}
\name{output_help}
\alias{output_help}
\title{Cell-ID output description}
\format{
An object of class \code{list} of length 46.
}
\usage{
output_help
}
\description{
Cell-ID output description
}
\keyword{datasets}
