% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\docType{data}
\name{output_help_df}
\alias{output_help_df}
\title{Cell-ID output description}
\format{
An object of class \code{data.frame} with 46 rows and 2 columns.
}
\usage{
output_help_df
}
\description{
Cell-ID output description
}
\keyword{datasets}
