% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{parameters_default}
\alias{parameters_default}
\title{Default parameters list for Cell-ID}
\usage{
parameters_default(
  max_split_over_minor = 0.5,
  max_dist_over_waist = 8,
  max_pixels_per_cell = 2000,
  min_pixels_per_cell = 75,
  background_reject_factor = 0.75,
  tracking_comparison = 0.2,
  align_individual_cells = F,
  align_fl_to_bf = F,
  align_fl_to_first = F,
  image_type = "brightfield",
  bf_fl_mapping = "list",
  treat_brightfield_as_fluorescence_also = F,
  third_image = F
)
}
\arguments{
\item{max_split_over_minor}{Default: \code{0.50} For every combination of two pixels on the boundary, Cell-ID calculates the distance along the boundary path divided by the Euclidean distance between them. The maximum value of this ratio is larger for cells with a “figure-eight” shape that were pinched in some part than for circular cells. If the maximum value is above a user-defined threshold (which defaults to max_dist_over_waist=6), then the cell is split into two cells at the location of the pinch. After a split, if the Euclidean distance divided by the length of the minor axis of either of the new cells is greater than a user-defined value (which defaults to max_split_over_minor=0.5), then the two cells are re-grouped as a single cell. Thus, to perform the split we require that the two new cells have a generally circular shape and are not too elongated, as would be the case if the previous split was not over two cells, but over a cell and its mating projection.}

\item{max_dist_over_waist}{Default: \code{8.00} For every combination of two pixels on the boundary, Cell-ID calculates the distance along the boundary path divided by the Euclidean distance between them. The maximum value of this ratio is larger for cells with a “figure-eight” shape that were pinched in some part than for circular cells. If the maximum value is above a user-defined threshold (which defaults to max_dist_over_waist=6), then the cell is split into two cells at the location of the pinch. After a split, if the Euclidean distance divided by the length of the minor axis of either of the new cells is greater than a user-defined value (which defaults to max_split_over_minor=0.5), then the two cells are re-grouped as a single cell. Thus, to perform the split we require that the two new cells have a generally circular shape and are not too elongated, as would be the case if the previous split was not over two cells, but over a cell and its mating projection.}

\item{max_pixels_per_cell}{Default: \code{2000} Area limits per cell (upper bound, in pixels).}

\item{min_pixels_per_cell}{Default: \code{75} Area limits per cell (lower bound, in pixels).}

\item{background_reject_factor}{Default: \code{0.75} CellID's code makes an initial decision about the graylevels of the boundary pixels. To do this it takes the mean position of all the graylevels in the images and subtracts Z standard deviations. It then starts by considering all gray levels below this value as being parts of the cell borders. This value Z is the parameter background_reject_factor. Brightfield images taken slightly out of focus may do better with with higher values (ie, higher values will better avoid spurious cells), but if the cell boundaries in the image are too narrow, a smaller value may be necessary--which might increase the level of background.}

\item{tracking_comparison}{Default: \code{0.20} Cell-ID attempts to track cells over time. The value of this parameter is the minimal fractional overlap between two cells in consecutive time points for them to be considered the same cell. The default value is 0.2. Also named \"I_over_U_for_match\" in CellID's cell.c and segment.c files.}

\item{align_individual_cells}{Default: \code{F} Allow wiggling between the brightfield and fluorescence images.}

\item{align_fl_to_bf}{Default: \code{F} Frame alignment. Cell-ID can perform image registrations, moving the the frame in XY to align it to a reference image. If “align FL to BF” is selected the bright field image is used as reference. If “align FL to first” is selected the first fluorescence image is used as reference. These options are especially useful when sampling different positions, as repositioning of the microscope stage might introduce some displacement between consecutive images.}

\item{align_fl_to_first}{Default: \code{F} To-do: document or link to explanation.}

\item{image_type}{Default: \code{"brightfield"} To-do: document or link to explanation.}

\item{bf_fl_mapping}{Default: \code{"list"} Possible values: "list", "time". "bf_fl_mapping" option description (guessed from code, mask_mod branch). The mapping between brightfield and fluorescence images can be made by acquisition time, or derived from the order in the list of paths passed as command line options "-b" and "-f" to cell. If the order is by "list", then the paths must be grouped and ordered first by t.frame (ascending) and then by channel. If the order is by "time", cell derives the BF-FL mapping from the acquisition time in the TIFF metadata.}

\item{treat_brightfield_as_fluorescence_also}{Default: \code{F} Calculate all the fluorescence images variables on the bright field image as if it were a fluorescence image. This is potentially a good idea since it allows a good way to reject spurious cells. For example, the average value of the boundary pixels in good cells will be lower than the background level, but not so for spurious cells, etc.}

\item{third_image}{Set to "vacuole_label" or "nuclear_label" to enable third image processing in Cell-ID. These images show the location of the nucleus or vacuole, and are used to derive measurements specific to those structures.}
}
\value{
A nice list of named parameters, input for \code{parameters_write}.
}
\description{
Returns a list of key-value pairs, for the default Cell-ID parameters.
It's output will tipically be used by \code{parameters_write} or the 
\code{arguments}.functions.
}
\details{
Documentation for each parameter can be found at: https://github.com/darksideoftheshmoo/cellID-linux#parameters

Boolean values are for "flag" type parameters
which enable a feature when present (eg. "align_fl_to_bf"),
or, if absent, indicate default behavior.

Other parameters have values
which must end up separated from names by a space " "
in the parameters.txt file format that Cell-ID uses:

\preformatted{
max_split_over_minor 0.5
max_dist_over_waist 8
max_pixels_per_cell 2000
min_pixels_per_cell 75
background_reject_factor 0.75
tracking_comparison 0.2
align_fl_to_bf
image_type brightfield
bf_fl_mapping list
}
}
\seealso{
\link[rcell2.cellid]{parameters_write}, \link[rcell2.cellid]{arguments}
}
