% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\docType{data}
\name{parameters_help}
\alias{parameters_help}
\title{Cell-ID parameter descriptions}
\format{
An object of class \code{list} of length 22.
}
\usage{
parameters_help
}
\description{
Cell-ID parameter descriptions
}
\keyword{datasets}
