% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\docType{data}
\name{parameters_help_df}
\alias{parameters_help_df}
\title{Cell-ID parameter descriptions}
\format{
An object of class \code{data.frame} with 23 rows and 3 columns.
}
\usage{
parameters_help_df
}
\description{
Cell-ID parameter descriptions
}
\keyword{datasets}
