% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{parameters_write}
\alias{parameters_write}
\title{Write parameters to a [temporary] file}
\usage{
parameters_write(
  parameters.list = rcell2.cellid::parameters_default(),
  param.dir = NULL,
  param.file = NULL
)
}
\arguments{
\item{parameters.list}{a parameters list for Cell-ID (like one from parameters_default)}

\item{param.dir}{directory where parameter files will be written.}

\item{param.file}{a file name for the parameters file.}
}
\value{
A path to the text file where parameters where written.
}
\description{
Parses a \code{parameters.list} object from \code{parameters_default},
and writes its contents to a Cell-ID friendly plain text file.
}
\seealso{
\link[rcell2.cellid]{parameters_default}, \link[rcell2.cellid]{arguments}
}
