% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiff_info.R
\name{parse_images_ch}
\alias{parse_images_ch}
\title{Parse images}
\usage{
parse_images_ch(image_list = NULL, images = NULL, channels = "BF")
}
\arguments{
\item{image_list}{The output from the \code{arguments} function. Consider filtering to include only the first position.}

\item{images}{The images dataframe as loaded from Cell-ID's output by \code{get_cell_data}.}

\item{channels}{Vector of channel IDs (e.g. "BF").}
}
\description{
Generate an "images" data frame, and filter it by channels. Also performs some checks.
}
