% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{rename_mda}
\alias{rename_mda}
\title{Image file renamer for Multi-dimensional acquisition image sets}
\usage{
rename_mda(
  images.path = NULL,
  rename.path = NULL,
  rename.dataframe = NULL,
  rename.function = file.symlink,
  identifier.pattern = ".*_w(\\\\d).*_s(\\\\d{1,2})_t(\\\\d{1,2}).TIF$",
  identifier.info = c("ch", Position = "pos", time = "t.frame"),
  channel_ids = 1:3,
  channel_names = c("CH1", "CH2", "CH3"),
  file.ext = ".tif",
  skip.thumbs.pat = ".*thumb.*",
  cleanup.first = FALSE,
  file.names = NULL,
  ...
)
}
\arguments{
\item{images.path}{Path to the directory containing the original images. Can be NULL if \code{file.names} is provided.}

\item{rename.path}{Path to the target directory. If \code{NULL} (the default) images are sent to a new "renamed" sub directory of \code{images.path}. The directory will be created if it does not exist.}

\item{rename.dataframe}{If \code{TRUE}, no renaming takes place and a data frame is returned instead. It can be altered by the user, and then passed to this same argument to rename the images with the adjusted names or mappings. The data frame is normally returned in the "status" item of the regular output.}

\item{rename.function}{Either \code{\link[base]{file.copy}}, \link[base]{file.symlink} or a similar function. Set to \code{NULL} to disable renaming (i.e. for testing purposes).}

\item{identifier.pattern}{Regex defining the iamge file pattern, with gropus for identifier in the file names.}

\item{identifier.info}{Character vector with strings "pos", "t.frame", and "ch" (channel), in the same order in which they appear in the \code{identifier.pattern}. If an element in the vector is named, the name is prefixed to the identifier in the final file name (for example, by default, "Position" is prepended to the position number; but channel has no prefix).}

\item{channel_ids}{Channel names in the source files (coerced to character).}

\item{channel_names}{New channel names for each channel in \code{channel_ids}.}

\item{file.ext}{File extension to use in the final file name, such as: ".tif".}

\item{skip.thumbs.pat}{A regex pattern to filter out files. Convenient if the MDA output thumbnails for each image. Set to \code{NULL} to disable.}

\item{cleanup.first}{Set to \code{TRUE} to remove all files within the \code{rename.path} directory before renaming. \code{FALSE} by default.}

\item{file.names}{A character vector of image names to be processed; as an alternative to listing files in \code{images.path}. Ignored if \code{images.path} is provided.}

\item{...}{Further arguments passed on to \code{rename.function}.}
}
\value{
Invisibly returns a list with the unique rename.paths (output directories), and a \code{data.frame} with the mappings for name conversions and the output value from the renaming function (see the \code{rename.function} parameter's description).
}
\description{
MDA: "Multi dimensional acquisition" app in Metamorph/MicroManager.
}
\details{
Uses regex groups to extract channel, position and time information from file names, and uses it to stitch new and friendlyer names.
These are used to copy or link image files to a target directory.

For example, \code{far1_rtcc_exp16_thumb_w1LED-BF--YFPcube--cam_s17_t35.TIF} can be converted to \code{BF_Position17_time35.tif}.

The \code{identifier.pattern} is a key parameter. There must be three groups, one for each of the three information types: channel, position and time.
The defaults are useful for a file name such as "\code{far1_rtcc_exp16_thumb_w1LED-BF--YFPcube--cam_s17_t35.TIF}", in which the channel is identified by a "w", 
position by an "s", and time by a "t".

The order in which this information appears in the file name is specified in \code{identifier.info}. 
If you wish to add a prefix to each field in the final file name, name the elements in this vector. 
For example, the default \code{c("ch", Position="pos", time="t.frame")} indicates that channel has no prefix,
the "pos" field will be prefixed by "Position", and the "t.frame" field will be prefixed by "time".
Then, for example, a new file name could look like this: \code{BF_Position1_time3.tif}.

Channel names will be translated according to the rows in \code{channel.dict} (see the parameter's description).
These are easily adaptable to other use cases, for example you may change \code{channel.dict} to include more, less or other channels, in whatever order.
Note that the values in the \code{ch} column must exactly match the strings captured by the corresponding capture group in \code{identifier.pattern}.
For example, the channel in the original file names may be integers from 1 to 3, which are captured and matched with dplyr's left_join to the \code{channel.dict} data frame.
Then, the value in \code{ch.name} is used to build the final file name.

**Limitations**: In the original file names, the identifiers for each field can only be integers.

Example:
images.path <- "~/path/to/dataset/"
rename_mda(images.path, rename.function = file.copy)
}
