% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiff_info.R
\name{tiff_plane_info}
\alias{tiff_plane_info}
\title{Get "PlaneInfo" from a TIFF's tags (metadata)}
\usage{
tiff_plane_info(path, frames = 1, ...)
}
\arguments{
\item{path}{A string. The path to the tiff file to read.}

\item{frames}{Which frames do you want to read tags from. Default first frame
only. To read from the 2nd and 7th frames, use \code{frames = c(2, 7)}, to read
from all frames, use \code{frames = "all"}.}

\item{...}{
  Arguments passed on to \code{\link[XML:xmlToList]{XML::xmlToList}}
  \describe{
    \item{\code{node}}{the XML node or document to be converted to an R list.
   This can be an "internal" or C-level node (i.e. \code{\link[XML]{XMLInternalNode-class}})
   or a regular R-level node (either \code{\link[XML]{XMLNode-class}} or \code{XMLHashNode}).}
    \item{\code{addAttributes}}{a logical value which controls whether the attributes of an empty node
    are added to the }
    \item{\code{simplify}}{a logical value that controls whether we collapse
   the list to a vector if the elements all have a common compatible
   type. Basically, this controls whether we use \code{sapply} or \code{lapply}.
 }
  }}
}
\value{
A dataframe with "prop" values of the TIFF's PlaneInfo metadata.
}
\description{
This function parses the XML metadata in Metamprph's TIFF image files,
and retrieves the "plane info" elements, containing information on
position, exposure, etc.

Only the first frame's information will be shown. Use this function once per frame.

This function requires:

* xmlToList from the XML package.

* read_tags from the ijtiff package.
}
