% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_utils.R
\docType{data}
\name{type_to_prototype}
\alias{type_to_prototype}
\title{List of prototypes in hierarchical order}
\format{
An object of class \code{list} of length 5.
}
\usage{
type_to_prototype
}
\description{
List of prototypes in hierarchical order
}
\keyword{datasets}
