% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{arguments_check}
\alias{arguments_check}
\title{Check if output files have already been created}
\usage{
arguments_check(arguments, check_fail = F)
}
\arguments{
\item{check_fail}{Whether to produce an error if any output file or directory already exists.}
}
\value{
TRUE if no existing output files were found
}
\description{
Emits warnings, and optionally an error, to prevent accidental overwriting.
}
