% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cell2}
\alias{cell2}
\title{Function to run Cell-ID}
\usage{
cell2(
  arguments,
  cell.command = NULL,
  n_cores = NULL,
  debug_flag = 0,
  dry = F,
  output_coords_to_tsv = F,
  encode_cellID_in_pixels = F,
  label_cells_in_bf = F,
  fill_interior_pixels = F,
  interior_offset = F,
  write_initial_time = F,
  save.logs = T,
  verbose = T,
  progress = F,
  check_fail = F
)
}
\arguments{
\item{arguments}{An argument data.frame, as built by \code{rcell2.cellid::arguments}.}

\item{cell.command}{By default \code{NULL}, to use the built-in binary. Otherwise a path to a Cell-ID binary executable (get if from https://github.com/darksideoftheshmoo/cellID-linux).}

\item{n_cores}{Number of cores to use for position-wise parallelization,internally capped to number of positions in \code{arguments}. Set to 1 to disable parallelization. If NULL, defaults to available cores - 1.}

\item{debug_flag}{Set to 0 to disable Cell-ID printf messages (built-in Cell-ID only).}

\item{dry}{Do everything without actually running Cell-ID, print the commands that would have been issued.}

\item{output_coords_to_tsv}{Set to TRUE to write cell interior and boundary pixels coordinates of each cell to a compressed \code{.tsv} file, located in the main output directory (Cell-ID option '-t'). This data can be loaded with the \code{cell.load.boundaries} function.}

\item{encode_cellID_in_pixels}{Set to TRUE to encode cellIDs in the intensity values of the boundary pixels, and blank the rest of the output image (Cell-ID option '-m'). Pixel intensities are proportional to each cellID, following the relationship \code{cellID = 65535 - boundary_intensity - 1}. Only boundary pixels are used by default; this behavior can be modified by enabling \code{label_cells_in_bf}, \code{fill_interior_pixels}, or \code{interior_offset}.}

\item{label_cells_in_bf}{Set mask boundary pixel intensities proportional to each cellID, and add cellID numbers to the cells with maximum pixel intensity \code{65535} (Cell-ID option '-l', default FALSE).}

\item{fill_interior_pixels}{Fill each cell interior area in the output BF image file with intensity-labeled pixels (Cell-ID option '-i'). This overrides cell labeling.}

\item{interior_offset}{Offset boundary and interior pixel intensities by a calculated 'interior_offset' threshold. cellID will relate to interior pixel intensities with the relationship \code{cellID = 65535 - boundary_intensity - interior_offset - 1}. The offset defaults to 5000, but may have a larger value for images or time series with more than 2500 cells (Cell-ID option '-w').}

\item{write_initial_time}{Write the absolute time of the first image to a text file (Cell-ID option '-z').}

\item{save.logs}{Set to TRUE to save Cell-ID logs to text files, into the output directory of their corresponding position.}

\item{verbose}{Print start-up messages.}

\item{progress}{Print a progress bar. Requires the \code{doSNOW} package.}

\item{check_fail}{Whether to produce an error if any output file or directory already exists.}
}
\value{
A data.frame with one column indicating the issued commands and exit codes (in the command.output column). If the execution was successful, now you may run \code{rcell2::load_cell_data} or \code{rcell2.cellid::get_cell_data} to get the results from the Cell-ID output, typically located at the images path.
}
\description{
Handles Cell-ID executions: creates necessary directories, sets-up parallelization, wraps the Cell-ID binary, and checks exit status.
}
