% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{check_and_fix_paths}
\alias{check_and_fix_paths}
\title{Check and fix image paths}
\usage{
check_and_fix_paths(path, images)
}
\arguments{
\item{path}{Path to the data directory, holding the images.}

\item{images}{Cell-ID dataframe with image paths, as loaded by \code{get_cell_data}.}
}
\description{
Check and fix image paths
}
