% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_utils.R
\name{get_spos_template}
\alias{get_spos_template}
\title{A function to copy the spos XLSX spreadsheet template to the current working directory.}
\usage{
get_spos_template(save_to = "stage_positions.xlsx", overwrite = FALSE)
}
\arguments{
\item{save_to}{File name for the downloaded workflow template.}

\item{overwrite}{Flag to overwrite the existing file if found.}
}
\description{
The template file is included in \code{cell2.cellid}.
}
