% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_template.R
\name{get_workflow_template_cellid}
\alias{get_workflow_template_cellid}
\title{A function to download the latest workflow template in Rmarkdown}
\usage{
get_workflow_template_cellid(
  file_name = "rcell2.cellid_workflow_template.Rmd",
  open.template = T
)
}
\arguments{
\item{file_name}{File name for the downloaded workflow template.}

\item{open.template}{Try using file.edit to open the file in RStudio after getting it.}
}
\description{
Will download the .Rmd file to the current working directory.
}
