% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_utils.R
\name{grid_to_long}
\alias{grid_to_long}
\title{Parse a grid sheet into long format}
\usage{
grid_to_long(sheet_data, rows_to = "row", names_to = "col", values_to = "row")
}
\description{
Parse a grid sheet into long format
}
