% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_macros.R
\name{ijm_open_hyperstack}
\alias{ijm_open_hyperstack}
\title{Make an ImageJ macro to browse the images before segmentation}
\usage{
ijm_open_hyperstack(images, use_out = FALSE, macro_file = TRUE)
}
\arguments{
\item{images}{The "images" dataframe, as loaded by cell.load.alt, load_cell_data, or as generated by arguments_to_images from arguments.}

\item{use_out}{Wether to look for original images, or the segmented output.}

\item{macro_file}{TRUE: write to temporary file/directory. FALSE: do not write macro to a file. Else, a path to the file where to save the macro.}
}
\description{
Due to an order restriction on ImageJ Virtual Hyperstacks, t.frame is mapped to "Channel", position is mapped to "Slice", and channel is mapped to "frames".
}
\details{
The function writes a macro to a temp
}
