% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_macros.R
\name{ijm_open_segmentation}
\alias{ijm_open_segmentation}
\title{ImageJ macro to open ".out" files as a virtual hyperstack}
\usage{
ijm_open_segmentation(cellid.args, write_macro = T)
}
\arguments{
\item{cellid.args}{The "arguments" data frame, as produced by \code{rcell2.cellid::arguments}.}

\item{write_macro}{Write the generated macro to an ".ijm" text file at the image's path.}
}
\description{
ImageJ macro to open ".out" files as a virtual hyperstack
}
