% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_parameter_scan.R
\name{make_scan_stacks}
\alias{make_scan_stacks}
\title{Make TIFF stacks for reviewing the result of parameter scans in ImageJ}
\usage{
make_scan_stacks(
  results.bound,
  test.dir,
  stack.channels = "BF.out",
  annotation.font = "Hack"
)
}
\arguments{
\item{results.bound}{The result from \code{parameter_scan}.}

\item{test.dir}{Temporary directory into which stacks should be saved.}

\item{stack.channels}{Vector of channels which should be stacked.}

\item{annotation.font}{Font for the annotations. A monospaced font is recommended.}
}
\value{
A list with paths to the stacks, and
}
\description{
Make TIFF stacks for reviewing the result of parameter scans in ImageJ
}
