% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiff_info.R
\name{plot_pos_overlaps}
\alias{plot_pos_overlaps}
\title{Plot positions and field of view overlaps}
\usage{
plot_pos_overlaps(
  image_list,
  file.pattern = "^(BF|[A-Z]FP)_Position(\\\\d+)_time(\\\\d+).tif$",
  magnification = 40,
  ccd_pixel_size_microns = 6.45
)
}
\arguments{
\item{image_list}{The output from the \code{arguments} function. Consider fltering to include only the first position.}

\item{file.pattern}{Regex pattern for all tif files, with one group for each of \code{c("ch", "pos", "t.frame")} in \code{file.pattern.groups.order}. Uses \code{"^(BF|[A-Z]FP)_Position(\\d+)_time(\\d+).tif$"} by default. To omit time, use an empty group for the t.frame in the regex, for example: \code{"^(BF|[A-Z]FP)_Position(\\d+)().tif$"}.
To consider Z-stacks, use something like "^(BF|[A-Z]\\d+)_Position(\\d+)_time(\\d+).tif$"}
}
\description{
Plot positions and field of view overlaps
}
\examples{
image_list <- rcell2.cellid::arguments(metamorph_pics_dir, 
                                       file.pattern = "^(BF|[A-Z]FP)_Position(\\\\d+)_time(\\\\d+).tif$") |> 
  filter(t.frame == min(t.frame)) |> 
  plot_pos_overlaps(file.pattern = "^(BF|[A-Z]FP)_Position(\\\\d+)_time(\\\\d+).tif$")

}
