% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{rename_mdas}
\alias{rename_mdas}
\title{Convenience function combining join_mdas and rename_mda}
\usage{
rename_mdas(
  images.path,
  index_pattern = "^MDA(\\\\d+)_.*",
  print_plot = T,
  identifier.pattern = ".*_w(\\\\d).*_s(\\\\d{1,2})_t(\\\\d{1,2}).TIF$",
  identifier.info = c("ch", Position = "pos", time = "t.frame"),
  channel_ids = 1:4,
  channel_names = c("CH1", "CH2", "CH3", "CH4"),
  rename.function = file.symlink,
  ...
)
}
\arguments{
\item{images.path}{Path to the directory containing the original images. Can be NULL if \code{file.names} is provided.}

\item{index_pattern}{Regular expression with a single capturing group for the MDA group index. The index matched by the pattern must be an integer, and mus uniquely identify each MDA run.}

\item{print_plot}{Prints a plot that can help check visually if the final image set will be complete.}

\item{identifier.pattern}{Regex defining the iamge file pattern, with gropus for identifier in the file names.}

\item{identifier.info}{Character vector with strings "pos", "t.frame", and "ch" (channel), in the same order in which they appear in the \code{identifier.pattern}. If an element in the vector is named, the name is prefixed to the identifier in the final file name (for example, by default, "Position" is prepended to the position number; but channel has no prefix).}

\item{channel_ids}{Channel names in the source files (coerced to character).}

\item{channel_names}{New channel names for each channel in \code{channel_ids}.}

\item{rename.function}{Either \code{\link[base]{file.copy}}, \link[base]{file.symlink} or a similar function. Set to \code{NULL} to disable renaming (i.e. for testing purposes).}

\item{...}{Further arguments passed on to \code{rename.function}.}
}
\description{
Convenience function combining join_mdas and rename_mda
}
