% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fft_files.R
\name{run_fft_filter_on_bfs}
\alias{run_fft_filter_on_bfs}
\title{Run Imagej's FFT filter on BF files}
\usage{
run_fft_filter_on_bfs(
  data.dir,
  cellid.args,
  fft.subdirs.prefix = "bf_subset",
  filtered.bfs.dir.name = "fft_images_dataset",
  imagej.path = "~/Software/ImageJ/Fiji.app/ImageJ-linux64",
  n_cores = NULL
)
}
\arguments{
\item{data.dir}{Data directory of the original images, appropriately named.}

\item{cellid.args}{Arguments dataframe as produced by \code{rcell2.cellid::arguments} at \code{data.dir}.}

\item{fft.subdirs.prefix}{Name prefix for intermediate filtered BF files.}

\item{filtered.bfs.dir.name}{Name for the final output directory,}

\item{imagej.path}{Path to the ImageJ/FIJI binary.}

\item{n_cores}{Integer amout of processor cores to use.}
}
\description{
Use ImageJ's bandpass FFT filter on the defocused brightfield images. This is optional, but can help Cell-ID find cells and reduce background artifacts.
}
\details{
Note that several intermediate files and directories are created during execution, which can remain there if the execution is interrupted for any reason.

To save disk space, fluorescence images are linked to the final directory as symbolic links instead of being duplicated by copying. Your milage may vary on a "non-unix OS" (i.e. Windows).
}
