% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_utils.R
\name{write_stg}
\alias{write_stg}
\title{Write position data in STG format}
\usage{
write_stg(stage_coords, positions, stg_output_path, overwrite = FALSE)
}
\arguments{
\item{stage_coords}{Stage coordinates data frame, as produced by \code{make_stage_list}.}

\item{positions}{Positions table as generated by \code{make_positions}.}

\item{stg_output_path}{Path where the STG file will be saved. This file is meant to be imported with MetaMorph's MDA.}

\item{overwrite}{Wether to overwrite the file at `stg_output_path` if it already exists.}
}
\description{
Write position data in STG format
}
